requirements.txt
flask[async]
flask-cors
asyncpg
websockets
beautifulsoup4
selenium
hypercorn
python-dotenv
